fix: flush results upon close
