+++ .github/workflows/bench.yml
@@ -1,0 +1,21 @@
+---
+name: Go-Benchmarks
+on:
+  push:
+    branches: ["main"]
+permissions: read-all
+jobs:
+  bench:
+    runs-on: ubuntu-latest
+    steps:
+      - name: checkout
+        uses: actions/checkout@v4
+      - name: gobenchdata publish
+        uses: bobheadxi/gobenchdata@v1
+        with:
+          PRUNE_COUNT: 30
+          GO_TEST_FLAGS: -cpu 1,2
+          PUBLISH: true
+          PUBLISH_BRANCH: gh-pages
+        env:
+          GITHUB_TOKEN: ${{ secrets.ACCESS_TOKEN }}
