+++ README.md
@@ -1,5 +1,11 @@
 # batcher
 
+![CI](https://github.com/NSXBet/batcher/actions/workflows/go.yml/badge.svg)
+![CI](https://github.com/NSXBet/batcher/actions/workflows/codeql.yml/badge.svg)
+[![Go Report Card](https://goreportcard.com/badge/github.com/NSXBet/batcher)](https://goreportcard.com/report/github.com/NSXBet/batcher)
+[![Maintainability](https://api.codeclimate.com/v1/badges/868870a2b4f7f29512ad/maintainability)](https://codeclimate.com/github/NSXBet/batcher/maintainability)
+[![Test Coverage](https://api.codeclimate.com/v1/badges/868870a2b4f7f29512ad/test_coverage)](https://codeclimate.com/github/NSXBet/batcher/test_coverage)
+
 The dead-simple batching solution for golang applications.
 
 With batcher you can easily batch operations and run them asynchronously in batches:
