+++ pkg/batcher/batcher_test.go
@@ -200,52 +200,33 @@ func TestCanHandleErrors(t *testing.T) {
 	require.NotNil(t, err)
 }
 
-func BenchmarkBatcherBatchSize10(b *testing.B) {
-	runBench(b, 10)
-}
+func TestFlushOnClose(t *testing.T) {
+	// ARRANGE
+	foundItems := sync.Map{}
+	b := batcher.New(
+		batcher.WithProcessor(func(items []test.BatchItem) error {
+			for _, item := range items {
+				foundItems.Store(item.Key, item)
+			}
 
-func BenchmarkBatcherBatchSize100(b *testing.B) {
-	runBench(b, 100)
-}
-
-func BenchmarkBatcherBatchSize1_000(b *testing.B) {
-	runBench(b, 1000)
-}
-
-func BenchmarkBatcherBatchSize10_000(b *testing.B) {
-	runBench(b, 100)
-}
-
-func BenchmarkBatcherBatchSize100_000(b *testing.B) {
-	runBench(b, 100)
-}
-
-func runBench(b *testing.B, batchSize int) {
-	b.StopTimer()
-
-	batch := batcher.New(
-		batcher.WithProcessor(func(_ []test.BatchItem) error {
 			return nil
 		}),
-		batcher.WithBatchSize[test.BatchItem](batchSize),
-		batcher.WithBatchInterval[test.BatchItem](1*time.Second),
+		batcher.WithBatchSize[test.BatchItem](100),
+		batcher.WithBatchInterval[test.BatchItem](1*time.Millisecond),
 	)
 
-	defer batch.Close()
-
-	b.StartTimer()
-
-	for i := 0; i < b.N; i++ {
-		batch.Add(test.BatchItem{Key: fmt.Sprintf("key_%d", i)})
+	// ACT
+	for i := 0; i < 1000; i++ {
+		b.Add(test.BatchItem{Key: fmt.Sprintf("key_%d", i)})
 	}
 
-	b.StopTimer()
+	// ASSERT
+	b.Close()
 
-	if err := batch.Join(5 * time.Second); err != nil {
-		b.Fatalf("error: %v", err)
-	}
+	require.Equal(t, 0, b.Len())
 
-	if batch.Len() > 0 {
-		b.Fatalf("expected 0 items in batch, got %d", batch.Len())
+	for i := 0; i < 1000; i++ {
+		_, ok := foundItems.Load(fmt.Sprintf("key_%d", i))
+		require.True(t, ok)
 	}
 }
